+++ spheres.go
@@ -66,7 +66,7 @@ func (s *Sphere) Intersections(r *Ray) []*Intersection {
 		return i
 	}
 
-	discriminantRoot := (float32)(math.Sqrt((float64)(discriminant)))
+	discriminantRoot := math.Sqrt(discriminant)
 	t1 := (-1*b - discriminantRoot) / (2 * a)
 	t2 := (-1*b + discriminantRoot) / (2 * a)
